+++ tests/test_trainer_models.py
@@ -1,203 +1,252 @@
-import os
 import pytest
-import shutil
 import torch
-from datetime import datetime
-import unittest
-import sys
-from _setup_local import config
-sys.path.append("../")
+from _setup_local import config, Net
 from dorc.device import all_devices, useable_devices
 from dorc.trainer import Trainer
+from util import SubTrainer
 
 
-class SubTrainer(Trainer):
-    def __init__(self, _cuda, *args, **kwargs):
-        self._cuda = _cuda
-        super().__init__(*args, **kwargs)
+class Net2(torch.nn.Module):
+    def __init__(fles):
+        super().__init__()
+        fles.a = Net()
+        fles.b = Net()
 
-    @property
-    def have_cuda(self):
-        return self._cuda
+    def forward(fles, x):
+        x_a = fles.a(x)
+        x_b = fles.b(x)
+        return x_a + x_b
 
 
 @pytest.mark.quick
-class TrainerTestModels(unittest.TestCase):
-    @classmethod
-    def setUpClass(cls):
-        """Setup a simple trainer with MNIST dataset."""
-        import importlib
-        import _setup_local as setup
-        importlib.reload(setup)
-        config = setup.config
-        cls.config = config
-        if os.path.exists(".test_dir"):
-            shutil.rmtree(".test_dir")
-        os.mkdir(".test_dir")
-        os.mkdir(".test_dir/global_modules")
-        os.mkdir(".test_dir/global_datasets")
-        os.mkdir(".test_dir/test_session")
-        time_str = datetime.now().isoformat()
-        os.mkdir(f".test_dir/test_session/{time_str}")
-        cls.data_dir = f".test_dir/test_session/{time_str}"
-        cls.params = {"global_modules_dir": ".test_dir/global_modules",
-                      "global_datasets_dir": ".test_dir/global_datasets",
-                      "data_dir": cls.data_dir, **cls.config}
-        cls.Net = cls.config["model_params"]["net"]["model"]
-
-    def test_trainer_init_models_one_model_no_gpus(self):
-        with self.subTest(i="no_gpus_trainer_params"):
-            self.params["model_params"] = {"net": {"model": self.Net,
-                                                   "optimizer": "Adam",
-                                                   "params": {},
-                                                   "gpus": []}}
-            self.trainer = SubTrainer(False, **self.params)
-            self.trainer.reserved_gpus = []
-            self.trainer.reserve_gpus = lambda x: [True, None]
-            self.trainer._init_device()
-            self.trainer._init_models()
-            self.assertIn("net", self.trainer._models)
-            self.assertEqual(self.trainer._models["net"].gpus, [])
-        with self.subTest(i="gpus_in_trainer_params"):
-            self.params["trainer_params"]["gpus"] = [0]
-            self.trainer = SubTrainer(False, **self.params)
-            self.trainer.reserved_gpus = []
-            self.trainer.reserve_gpus = lambda x: [True, None]
-            self.trainer._init_device()
-            self.trainer._init_models()
-            self.assertIn("net", self.trainer._models)
-            self.assertEqual(self.trainer._models["net"].gpus, [])
-
-    @unittest.skipIf(not all_devices(), f"Cannot run without GPUs.")
-    def test_trainer_init_models_one_model_one_gpu(self):
-        self.params["model_params"] = {"net": {"model": self.Net,
-                                               "optimizer": "Adam",
-                                               "params": {},
-                                               "gpus": [0]}}
-        self.params["trainer_params"]["gpus"] = [0]
-        self.trainer = SubTrainer(False, **self.params)
-        self.trainer.reserved_gpus = []
-        self.trainer.reserve_gpus = lambda x: [True, None]
-        self.trainer.trainer_params.cuda = True
-        self.trainer._init_device()
-        self.trainer._init_models()
-        self.assertIn("net", self.trainer._models)
-        self.assertEqual(self.trainer._models["net"].gpus, [0])
-
-    @unittest.skipIf(not all_devices(), f"Cannot run without GPUs.")
-    def test_trainer_init_models_many_models_no_auto_no_parallel_gpus_sufficient(self):
-        self.params["model_params"] = {"net_1": {"model": self.Net,
-                                                 "optimizer": "Adam",
-                                                 "params": {},
-                                                 "gpus": [0]},
-                                       "net_2": {"model": self.Net,
-                                                 "optimizer": "Adam",
-                                                 "params": {},
-                                                 "gpus": [1]}}
-        self.params["trainer_params"]["gpus"] = [0, 1]
-        self.trainer = SubTrainer(False, **self.params)
-        self.trainer.reserved_gpus = []
-        self.trainer.reserve_gpus = lambda x: [True, None]
-        self.trainer.trainer_params.cuda = True
-        self.trainer._init_device()
-        self.trainer._init_models()
-        self.assertIn("net_1", self.trainer._models)
-        self.assertIn("net_2", self.trainer._models)
-        self.assertEqual(self.trainer._models["net_1"].gpus, [0])
-        self.assertEqual(self.trainer._models["net_2"].gpus, [1])
-
-    @unittest.skipIf(not all_devices(), f"Cannot run without GPUs.")
-    def test_trainer_init_models_many_models_no_auto_no_parallel_gpus_deficient(self):
-        self.params["model_params"] = {"net_1": {"model": self.Net,
-                                                 "optimizer": "Adam",
-                                                 "params": {},
-                                                 "gpus": [0, 1]},
-                                       "net_2": {"model": self.Net,
-                                                 "optimizer": "Adam",
-                                                 "params": {},
-                                                 "gpus": [2, 3]}}
-        self.params["trainer_params"]["gpus"] = [0, 1]
-        self.trainer = SubTrainer(False, **self.params)
-        self.trainer.reserved_gpus = []
-        self.trainer.reserve_gpus = lambda x: [True, None]
-        self.trainer.trainer_params.cuda = True
-        self.trainer._init_device()
-        self.trainer._init_models()
-        self.assertIn("net_1", self.trainer._models)
-        self.assertIn("net_2", self.trainer._models)
-        self.assertEqual(self.trainer._models["net_1"].gpus, [0, 1])
-        self.assertEqual(self.trainer._models["net_2"].gpus, [])
-
-    @unittest.skipIf(not all_devices(), f"Cannot run without GPUs.")
-    def test_trainer_init_models_many_models_no_auto_no_parallel_gpus_conflict(self):
-        self.params["model_params"] = {"net_1": {"model": self.Net,
-                                                 "optimizer": "Adam",
-                                                 "params": {},
-                                                 "gpus": [0, 1]},
-                                       "net_2": {"model": self.Net,
-                                                 "optimizer": "Adam",
-                                                 "params": {},
-                                                 "gpus": [0]}}
-        self.params["trainer_params"]["gpus"] = [0, 1]
-        self.trainer = SubTrainer(False, **self.params)
-        self.trainer.reserved_gpus = []
-        self.trainer.reserve_gpus = lambda x: [True, None]
-        self.trainer.trainer_params.cuda = True
-        self.trainer._init_device()
-        self.trainer._init_models()
-        self.assertIn("net_1", self.trainer._models)
-        self.assertIn("net_2", self.trainer._models)
-        self.assertEqual(self.trainer._models["net_1"].gpus, [0, 1])
-        self.assertEqual(self.trainer._models["net_2"].gpus, [])
-
-    @unittest.skipIf(len(all_devices()) < 2, f"Cannot run without at least 2 GPUs.")
-    def test_trainer_init_models_two_models_two_gpus_auto_no_parallel(self):
-        bleh = self.Net
-
-        class Net2(torch.nn.Module):
-            def __init__(fles):
-                super().__init__()
-                fles.a = bleh()
-                fles.b = bleh()
-
-            def forward(fles, x):
-                x_a = fles.a(x)
-                x_b = fles.b(x)
-                return x_a + x_b
-
-        self.params["model_params"] = {"net_1": {"model": self.Net,
-                                                 "optimizer": "Adam",
-                                                 "params": {},
-                                                 "gpus": "auto"},
-                                       "net_2": {"model": Net2,
-                                                 "optimizer": "Adam",
-                                                 "params": {},
-                                                 "gpus": "auto"}}
-        self.params["trainer_params"]["gpus"] = [0, 1]
-        self.trainer = SubTrainer(False, **self.params)
-        self.trainer.reserved_gpus = []
-        self.trainer.reserve_gpus = lambda x: [True, None]
-        self.trainer.trainer_params.cuda = True
-        self.trainer._init_device()
-        self.trainer._init_models()
-        self.assertIn("net_1", self.trainer._models)
-        self.assertIn("net_2", self.trainer._models)
-        self.assertEqual(self.trainer._models["net_1"].gpus, [1])
-        self.assertEqual(self.trainer._models["net_2"].gpus, [0])
-
-    # TODO
-    def test_trainer_load_models_state(self):
-        pass
-
-    # TODO
-    def test_trainer_get_new_models(self):
-        pass
-
-    @classmethod
-    def tearDownClass(cls):
-        if os.path.exists(".test_dir"):
-            shutil.rmtree(".test_dir")
+def test_trainer_init_models_one_model_no_gpus_and_no_gpus_in_params_inits_correctly(params_and_trainer):
+    params, trainer = params_and_trainer
+    params["model_params"] = {"net": {"model": Net,
+                                      "optimizer": "Adam",
+                                      "params": {},
+                                      "gpus": []}}
+    trainer = SubTrainer(False, **params)
+    trainer._init_device()
+    trainer._init_models()
+    trainer._init_data_and_dataloaders()
+    trainer._init_update_funcs()
+    trainer._init_training_steps()
+    assert "net" in trainer._training_steps["train"][0].models
+    if trainer._model_step_func:
+        assert trainer.active_models == {}
+        trainer.set_model({"net", "net"})
+        assert trainer._models["net"].gpus == []
+    else:
+        assert trainer._training_steps["train"][0].models['net'].loaded
+        assert trainer.active_models == {"net": "net"}
+        assert trainer._models["net"].gpus == []
 
 
-if __name__ == '__main__':
-    unittest.main()
+@pytest.mark.quick
+@pytest.mark.parametrize("params_and_trainer", [True, False], indirect=True)
+def test_trainer_init_models_one_model_no_gpus_but_gpus_in_params_init_correctly(params_and_trainer):
+    params, trainer = params_and_trainer
+    params["trainer_params"]["gpus"] = [0]
+    trainer = SubTrainer(False, **params)
+    trainer._init_device()
+    trainer._init_models()
+    trainer._init_data_and_dataloaders()
+    trainer._init_update_funcs()
+    trainer._init_training_steps()
+    assert "net" in trainer._training_steps["train"][0].models
+    if trainer._model_step_func:
+        assert trainer.active_models == {}
+    else:
+        assert trainer._training_steps["train"][0].models['net'].loaded
+        assert trainer.active_models == {"net": "net"}
+        assert trainer._models["net"].gpus == []
+
+
+@pytest.mark.quick
+@pytest.mark.skipif(not all_devices(), reason=f"Cannot run without GPUs.")
+@pytest.mark.parametrize("params_and_trainer", [True, False], indirect=True)
+def test_trainer_init_models_one_model_one_gpu(params_and_trainer):
+    params, trainer = params_and_trainer
+    params["model_params"] = {"net": {"model": Net,
+                                      "optimizer": "Adam",
+                                      "params": {},
+                                      "gpus": [0]}}
+    params["trainer_params"]["gpus"] = [0]
+    trainer = SubTrainer(False, **params)
+    trainer.trainer_params.cuda = True
+    trainer._init_device()
+    trainer._init_models()
+    trainer._init_data_and_dataloaders()
+    trainer._init_update_funcs()
+    trainer._init_training_steps()
+    assert "net" in trainer._traini_steps["train"][0].models
+    if trainer._model_step_func:
+        assert trainer.active_models == {}
+    else:
+        assert trainer._training_steps["train"][0].models['net'].loaded
+        assert trainer.active_models == {"net": "net"}
+        assert trainer._models["net"].gpus == [0]
+
+
+@pytest.mark.skipif(not all_devices(), reason=f"Cannot run without GPUs.")
+@pytest.mark.quick
+def test_trainer_init_models_many_models_no_auto_no_parallel_gpus_sufficient(params_and_trainer):
+    params, trainer = params_and_trainer
+    params["model_params"] = {"net_1": {"model": Net,
+                                        "optimizer": "Adam",
+                                        "params": {},
+                                        "gpus": [0]},
+                              "net_2": {"model": Net,
+                                        "optimizer": "Adam",
+                                        "params": {},
+                                        "gpus": [1]}}
+    params["trainer_params"]["gpus"] = [0, 1]
+    trainer = SubTrainer(False, **params)
+    trainer.trainer_params.cuda = True
+    trainer._init_device()
+    trainer._init_models()
+    trainer._init_data_and_dataloaders()
+    trainer._init_update_funcs()
+    trainer._init_training_steps()
+    assert "net_1" in trainer._models
+    assert "net_2" in trainer._models
+    assert trainer._models["net_1"].gpus == [0]
+    assert trainer._models["net_2"].gpus == [1]
+
+
+@pytest.mark.skipif(not all_devices(), reason=f"Cannot run without GPUs.")
+@pytest.mark.quick
+def test_trainer_init_models_many_models_no_auto_no_parallel_gpus_deficient(params_and_trainer):
+    params, trainer = params_and_trainer
+    params["model_params"] = {"net_1": {"model": Net,
+                                        "optimizer": "Adam",
+                                        "params": {},
+                                        "gpus": [0, 1]},
+                              "net_2": {"model": Net,
+                                        "optimizer": "Adam",
+                                        "params": {},
+                                        "gpus": [2, 3]}}
+    params["trainer_params"]["gpus"] = [0, 1]
+    trainer = SubTrainer(False, **params)
+    trainer.trainer_params.cuda = True
+    trainer._init_device()
+    trainer._init_models()
+    trainer._init_data_and_dataloaders()
+    trainer._init_update_funcs()
+    trainer._init_training_steps()
+    assert "net_1" in trainer._models
+    assert "net_2" in trainer._models
+    assert trainer._models["net_1"].gpus == [0, 1]
+    assert trainer._models["net_2"].gpus == []
+
+
+@pytest.mark.skipif(not all_devices(), reason=f"Cannot run without GPUs.")
+@pytest.mark.quick
+def test_trainer_init_models_many_models_no_auto_no_parallel_gpus_conflict(params_and_trainer):
+    params, trainer = params_and_trainer
+    params["model_params"] = {"net_1": {"model": Net,
+                                             "optimizer": "Adam",
+                                             "params": {},
+                                             "gpus": [0, 1]},
+                                   "net_2": {"model": Net,
+                                             "optimizer": "Adam",
+                                             "params": {},
+                                             "gpus": [0]}}
+    params["trainer_params"]["gpus"] = [0, 1]
+    trainer = SubTrainer(False, **params)
+    trainer.trainer_params.cuda = True
+    trainer._init_device()
+    trainer._init_models()
+    trainer._init_data_and_dataloaders()
+    trainer._init_update_funcs()
+    trainer._init_training_steps()
+    assert "net_1" in trainer._models
+    assert "net_2" in trainer._models
+    assert trainer._models["net_1"].gpus == [0, 1]
+    assert trainer._models["net_2"].gpus == []
+
+
+@pytest.mark.skipif(len(all_devices()) < 2, reason=f"Cannot run without at least 2 GPUs.")
+@pytest.mark.quick
+def test_trainer_init_models_two_models_two_gpus_auto_no_parallel(params_and_trainer):
+    params, trainer = params_and_trainer
+    params["model_params"] = {"net_1": {"model": Net,
+                                        "optimizer": "Adam",
+                                        "params": {},
+                                        "gpus": "auto"},
+                              "net_2": {"model": Net2,
+                                        "optimizer": "Adam",
+                                        "params": {},
+                                        "gpus": "auto"}}
+    params["trainer_params"]["gpus"] = [0, 1]
+    trainer = SubTrainer(False, **params)
+    trainer.trainer_params.cuda = True
+    trainer._init_device()
+    trainer._init_models()
+    trainer._init_data_and_dataloaders()
+    trainer._init_update_funcs()
+    trainer._init_training_steps()
+    assert "net_1" in trainer._models
+    assert "net_2" in trainer._models
+    assert trainer._models["net_1"].gpus == [1]
+    assert trainer._models["net_2"].gpus == [0]
+
+
+# TODO
+@pytest.mark.quick
+@pytest.mark.parametrize("params_and_trainer", [True, False], indirect=True)
+def test_trainer_set_model_active(params_and_trainer):
+    params, trainer = params_and_trainer
+    params["model_params"] = {"net_1": {"model": Net,
+                                        "optimizer": "Adam",
+                                        "params": {},
+                                        "gpus": "auto"},
+                              "net_2": {"model": Net2,
+                                        "optimizer": "Adam",
+                                        "params": {},
+                                        "gpus": "auto"}}
+    trainer = SubTrainer(False, **params)
+    trainer._init_all()
+    if trainer._model_step_func:
+        assert trainer.models_available == ["net_1", "net_2"]
+        assert trainer.active_models == {}
+    else:
+        assert trainer.models_available == ["net_1", "net_2", "net"]
+        assert trainer.active_models == {"net": "net"}
+    trainer.set_model({"net": "net_1"})
+    assert trainer.active_models == {"net": "net_1"}
+    trainer._training_steps["test"][0](trainer.train_loader.__iter__().__next__())
+
+
+# @pytest.mark.quick
+# def test_trainer_set_model_active_alt(trainer):
+#     trainer._init_all()
+#     trainer._training_steps["test"][0](trainer.train_loader.dataset[0])
+#     pytest.set_trace()
+
+
+# TODO
+@pytest.mark.quick
+def test_trainer_add_model(params_and_trainer):
+    params, trainer = params_and_trainer
+    pass
+
+
+# TODO
+@pytest.mark.quick
+def test_trainer_load_and_unload_model(params_and_trainer):
+    params, trainer = params_and_trainer
+    pass
+
+
+# TODO
+@pytest.mark.quick
+def test_trainer_load_models_state(params_and_trainer):
+    params, trainer = params_and_trainer
+    pass
+
+# TODO
+@pytest.mark.quick
+def test_trainer_get_new_models(params_and_trainer):
+    params, trainer = params_and_trainer
+    pass